% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtimer.R
\name{worldtimer-functions}
\alias{worldtimer}
\alias{worldtimer_timezones}
\alias{worldtimer_ip}
\title{Send Requests to WorldTimeAPI}
\usage{
worldtimer(resource, ...)

worldtimer_timezones()

worldtimer_ip(ipv4 = NULL)
}
\arguments{
\item{resource}{Path in the WorldTimeAPI. Normally either a valid timezone
or an IP.}

\item{...}{
  Arguments passed on to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json}}
  \describe{
    \item{\code{resp}}{A response object.}
    \item{\code{check_type}}{Check that response has expected content type? Set to
\code{FALSE} to suppress the automated check}
    \item{\code{simplifyVector}}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}
  }}

\item{ipv4}{An ipv4 address for which to return the current time. The default
\code{NULL} uses the system's public IP address.}
}
\value{
\code{list} of class \code{worldtimer}
}
\description{
Functions to talk to the WorldTimeAPI \url{http://worldtimeapi.org}
}
\examples{
\donttest{
# Change R option for printing decimal digits of seconds
old <- options(digits.secs = 6)

# Specific timezone
new_york_time <- try(worldtimer("America/New_York"))
str(new_york_time)
print(new_york_time)

# The system's public ip
system_ip <- worldtimer_ip()
str(system_ip)
print(system_ip)

# Other public ip
other_ip <- worldtimer_ip("101.110.34.62")
str(other_ip)
print(other_ip)

# Restore old options
options(old)
}
}
