% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtimer.R
\name{worldtimer-methods}
\alias{worldtimer-methods}
\alias{print.worldtimer}
\alias{format.worldtimer}
\title{Print and Format worldtimer Objects}
\usage{
\method{print}{worldtimer}(x, ...)

\method{format}{worldtimer}(x, tz = "UTC", usetz = TRUE, ...)
}
\arguments{
\item{x}{an object to be printed or summarized from one of the
    date-time classes.}

\item{...}{further arguments to be passed from or to other methods.}

\item{tz}{for timezone formatting, passed to \code{\link[base]{format.POSIXct}}.}

\item{usetz}{for timezone formatting, passed to \code{\link[base]{format.POSIXct}}.}
}
\description{
These are Methods for the \code{print()} and \code{format()} generics.
}
\seealso{
\link{worldtimer-functions}
}
