% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtimeR-package.R
\docType{package}
\name{worldtimer-package}
\alias{worldtimer}
\alias{worldtimer-package}
\title{worldtimer: An R Package to Obtain Time from WorldTimeAPI}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{internal}
