% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtimer.R
\name{worldtimer}
\alias{worldtimer}
\alias{worldtimer_timezones}
\alias{worldtimer_ip}
\title{Send Requests to WorldTimeAPI}
\usage{
worldtimer(resource, ...)

worldtimer_timezones()

worldtimer_ip(ipv4 = NULL)
}
\arguments{
\item{resource}{Path in the WorldTimeAPI. Normally either a valid timezone
or an IP.}

\item{...}{
  Arguments passed on to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json}}
  \describe{
    \item{\code{resp}}{A response object.}
    \item{\code{check_type}}{Check that response has expected content type? Set to
\code{FALSE} to suppress the automated check}
    \item{\code{simplifyVector}}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}
  }}
}
\description{
Functions to talk to the WorldTimeAPI \url{http://worldtimeapi.org}
}
\examples{
/donttest{
try(worldtimer("America/New_York"))
}
}
